viem scripts updated
