+++ package.json
@@ -106,7 +106,7 @@
     "notify-send": "^0.1.2",
     "pino": "^7.9.2",
     "tsx": "^4.7.2",
-    "viem": "^2.11.1",
+    "viem": "^2.18.7",
     "zx": "^8.0.2"
   },
   "lint-staged": {
