+++ script/deploy/safe/add-owners-to-safe.ts
@@ -1,25 +1,11 @@
 import { defineCommand, runMain } from 'citty'
 import { type SafeApiKitConfig } from '@safe-global/api-kit'
-import { type Chain, defineChain, getAddress } from 'viem'
-import Safe, {
-  ContractNetworksConfig,
-  EthersAdapter,
-} from '@safe-global/protocol-kit'
+import { getAddress } from 'viem'
+import Safe, { EthersAdapter } from '@safe-global/protocol-kit'
 import SafeApiKit from '@safe-global/api-kit'
 import { ethers } from 'ethers6'
-import * as chains from 'viem/chains'
-import {
-  chainNameMappings,
-  getSafeUtilityContracts,
-  safeAddresses,
-  safeApiUrls,
-} from './config'
-
-const chainMap: Record<string, Chain> = {}
-for (const [k, v] of Object.entries(chains)) {
-  // @ts-ignore
-  chainMap[k] = v
-}
+import { getSafeUtilityContracts, safeAddresses, safeApiUrls } from './config'
+import { getViemChainForNetworkName } from '../../../utils/viemScriptHelpers'
 
 const main = defineCommand({
   meta: {
@@ -48,47 +34,18 @@ const main = defineCommand({
     },
   },
   async run({ args }) {
-    // const chainName = chainNameMappings[args.network] || args.network
-    // const chain: Chain = chainMap[chainName]
+    const { network, privateKey } = args
 
-    const chainName = 'immutablezkevm'
-    const chain = defineChain({
-      id: 13371,
-      name: 'Immutable zkEVM',
-      nativeCurrency: {
-        decimals: 18,
-        name: 'IMX',
-        symbol: 'IMX',
-      },
-      rpcUrls: {
-        default: {
-          http: ['https://rpc.immutable.com'], // [pre-commit-checker: not a secret]
-          webSocket: ['wss://immutable-zkevm.drpc.org'],
-        },
-      },
-      blockExplorers: {
-        default: { name: 'Explorer', url: 'https://explorer.immutable.com/' },
-      },
-      contracts: {
-        // multicall3: {
-        //   address: '0x4f6F1CeE578C0B201c65CCaaBFF1D924377F622D',
-        //   blockCreated: 2863004,
-        // },
-        multicall3: {
-          address: '0xcA11bde05977b3631167028862bE2a173976CA11',
-          blockCreated: 3680945,
-        },
-      },
-    })
+    const chain = getViemChainForNetworkName(network)
 
     const config: SafeApiKitConfig = {
       chainId: BigInt(chain.id),
-      txServiceUrl: safeApiUrls[chainName.toLowerCase()],
+      txServiceUrl: safeApiUrls[network],
     }
 
     const safeService = new SafeApiKit(config)
 
-    const safeAddress = getAddress(safeAddresses[chainName.toLowerCase()])
+    const safeAddress = getAddress(safeAddresses[network])
 
     const rpcUrl = args.rpcUrl || chain.rpcUrls.default.http[0]
     const provider = new ethers.JsonRpcProvider(rpcUrl)
@@ -134,7 +91,6 @@ const main = defineCommand({
       console.info('Adding owner', owner)
       console.info('Signer Address', senderAddress)
       console.info('Safe Address', safeAddress)
-      console.info('Network', chainName)
 
       // Propose transaction to the service
       await safeService.proposeTransaction({
